requirements.txt
tensorflow>=2.11
scikit-learn
matplotlib
pandas
opencv-python
albumentations
dvc
mlflow
pyngrok
fastapi
uvicorn
pillow
numpy
python-multipart

